core: convert sync-callback-runner to typescript
