+++ packages/core/src/index.ts
@@ -7,5 +7,6 @@ export { default as cloneClient } from './lib/clone-client'
 export { default as jsonPayload } from './lib/json-payload'
 export { default as intRange } from './lib/validators/int-range'
 export { default as isError } from './lib/iserror'
+export { default as runSyncCallbacks } from './lib/sync-callback-runner'
 
 export * from './common'
